Extract lollipop to a separate view
